Vaguely working game logic:
 * New Trick class to wrap these SlotsMap better, and like AllBids.
    Has listener for first card, e.g. for setting the lead suit elsewhere
 * Improved card comparators to allow lead suits with trumps
 * Move more state into Game
 * Add more setup to GameSetup
 * TODO: actually follow suit (implement allowed cards)

+++ src/main/java/net/declension/games/cards/ohhell/Scorer.java
@@ -4,6 +4,6 @@ import net.declension.games.cards.tricks.BidAndTaken;
 
 import java.util.function.Function;
 
-interface Scorer extends Function<BidAndTaken, Short> {
+interface Scorer extends Function<BidAndTaken, Integer> {
 
 }
